Corrected template to pluralize entity name. (#242)
Correcting the template where it attempts to pluralize the entity name of the crud module.  The template was incorrectly concatenating an 's' onto the end of the singular entity name where the inflections module had actually mapped the plural correctly.  The correction essentially changes the template to use the pluralized versions of the entity name.  This issue is particularly important because it is reproducable by following the meanjs.org superhero / superheroes video tutorial on youtube.com.